med grafisk calc

+++ src/java/facades/UserFacade.java
@@ -3,6 +3,7 @@ package facades;
 import entity.Role;
 import security.IUserFacade;
 import entity.User;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -13,20 +14,26 @@ import openshift_deploy.DeploymentConfiguration;
 import security.IUser;
 import security.PasswordStorage;
 
-public class UserFacade implements IUserFacade {
+public class UserFacade implements IUserFacade
+{
 
     EntityManagerFactory emf = Persistence.createEntityManagerFactory(DeploymentConfiguration.PU_NAME);
 
-    public UserFacade() {
+    public UserFacade()
+    {
 
     }
 
     @Override
-    public IUser getUserByUserId(String id) {
+    public IUser getUserByUserId(String id)
+    {
         EntityManager em = emf.createEntityManager();
-        try {
+        try
+        {
             return em.find(User.class, id);
-        } finally {
+        }
+        finally
+        {
             em.close();
         }
     }
@@ -38,31 +45,42 @@ public class UserFacade implements IUserFacade {
     /*
      Return the Roles if users could be authenticated, otherwise null
      */
-    public List<String> authenticateUser(String userName, String password) {
+    public List<String> authenticateUser(String userName, String password)
+    {
         EntityManager em = emf.createEntityManager();
-        try {
+        try
+        {
             User user = em.find(User.class, userName);
-            if (user == null) {
+            if (user == null)
+            {
                 return null;
             }
 
             boolean authenticated;
-            try {
+            try
+            {
                 authenticated = PasswordStorage.verifyPassword(password, user.getPassword());
                 return authenticated ? user.getRolesAsStrings() : null;
-            } catch (PasswordStorage.CannotPerformOperationException | PasswordStorage.InvalidHashException ex) {
+            }
+            catch (PasswordStorage.CannotPerformOperationException | PasswordStorage.InvalidHashException ex)
+            {
                 Logger.getLogger(UserFacade.class.getName()).log(Level.SEVERE, null, ex);
                 return null;
             }
-        } finally {
+        }
+        finally
+        {
             em.close();
         }
     }
 
-    public void saveUser(User newUsr) {
+    public void saveUser(User newUsr)
+    {
         EntityManager em = emf.createEntityManager();
-        try {
+        try
+        {
             em.getTransaction().begin();
+//            newUsr.AddRole(DeploymentConfiguration.userRole);
             newUsr.AddRole(em.find(Role.class, DeploymentConfiguration.userRole.getRoleName()));
             System.out.println("Hejsa " + newUsr.getPassword() + " - " + newUsr.getUserName() + " - " + newUsr.getRolesAsStrings());
 
@@ -70,33 +88,61 @@ public class UserFacade implements IUserFacade {
 
             em.persist(newUsr);
             em.getTransaction().commit();
-        } catch (PasswordStorage.CannotPerformOperationException | org.eclipse.persistence.exceptions.DatabaseException ex) {
+        }
+        catch (PasswordStorage.CannotPerformOperationException | org.eclipse.persistence.exceptions.DatabaseException ex)
+        {
             Logger.getLogger(UserFacade.class.getName()).log(Level.SEVERE, null, ex);
-        } finally {
+        }
+        finally
+        {
             em.close();
         }
     }
 
-    Role Role = new Role();
-    List<User> users = Role.getUsers();
-
-    public void deleteUser(String name) {
-
+    public void deleteUser(String name)
+    {
+        System.out.println("VI er i delUSR");
         EntityManager em = emf.createEntityManager();
-        try {
-
-//            for (int i = 0; i < users.size(); i++) {
-//                if (users.get(i).getUserName().equals(name)) {
-//                    users.remove(i);
-//                }
-//            }
+        try
+        {
             em.getTransaction().begin();
             User p = em.find(User.class, name);
-            em.remove(p);
-            em.getTransaction().commit();
-        } finally {
+            
+            
+            if (p != null)
+            {
+                System.out.println("User er ikke null");
+                em.remove(p);
+                DeploymentConfiguration.userRole = em.find(entity.Role.class, DeploymentConfiguration.userRole.getRoleName());
+                em.getTransaction().commit();
+            }
+
+        }
+        finally
+        {
             em.close();
         }
+    }
+
+    public List<entity.User> getAllUsers()
+    {
+        EntityManager em = emf.createEntityManager();
+        try
+        {
+            em.getTransaction().begin();
+            entity.Role rolle = em.find(entity.Role.class, DeploymentConfiguration.userRole.getRoleName());
+            System.out.println("Her er list: " + rolle.getUsers().toString());
+            em.getTransaction().commit();
+            System.out.println("Her er rolle: " + rolle.toString());
+            
+            return rolle.getUsers();
+            
+        }
+        finally
+        {
+            em.close();
+        }
+
     }
 
 }
